Chase-colors of newly joined players
There was a problem that new players who joined LVP while a chase was
running still got a random color instead of white, since he is not being
chased. Fixed that in this issue for #34, although keep in mind that
when coming out of a minigame this player will still have an other color
then white.