Skip to previous or next by arrow left or right while /watch'ing
With this functionality when you press the left key /watch will be done
with 1 id lower, unless it is the lowest then it selects the highest. At
pressing the right key it will be done otherwise, if the highest id is
watched it will set it to 0.
It includes the same checks as onWatchCommand.