Restrict /house, /property, /p deathmessage and /v [delete, save] for temporary administrators
These commands have a lasting effect, and are required only rarely in regular gameplay -- let alone for people with temporary rights. We therefore restrict them to normal administrators and up.

In order to implement this, code has been added to both the Pawn and JavaScript portions of our gamemode to propagate whether a player's level is temporary or not. In practice we only deal with temporary administrators.

Tests have been added and updated where appropriate, to make sure that we have sufficient coverage for this.

Part of Issue #512.